# docs/public/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "fyndnote",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/datasets/{ds_id}/details": {
      "get": {
        "tags": [
          "Datasets"
        ],
        "summary": "Dataset Details",
        "operationId": "dataset_details_api_v1_datasets__ds_id__details_get",
        "parameters": [
          {
            "name": "ds_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Ds Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/datasets/load": {
      "post": {
        "tags": [
          "Datasets"
        ],
        "summary": "Load Dataset",
        "operationId": "load_dataset_api_v1_datasets_load_post",
        "parameters": [
          {
            "name": "user_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1,
              "title": "User Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "additionalProperties": true,
                "title": "Body"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/datasets/upload": {
      "post": {
        "tags": [
          "Datasets"
        ],
        "summary": "Upload Dataset",
        "description": "Stream an uploaded dataset to disk instead of buffering it in RAM.\n\nStarlette already spools multipart parts to disk past ~1 MB, so the payload\nnever has to become a single ``bytes`` object; the previous\n``content = await file.read()`` materialised the whole file (a 1.1 GB CSV\npeaked the worker at ~1.4 GB RSS and pinned it for the whole transfer).\nAn oversized ``Content-Length`` is refused by :class:`UploadSizeGuard`\nbefore the body is read at all; the size of the spooled part is then\nenforced while copying, so oversized files never reach pyarrow.",
        "operationId": "upload_dataset_api_v1_datasets_upload_post",
        "parameters": [
          {
            "name": "user_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1,
              "title": "User Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_upload_dataset_api_v1_datasets_upload_post"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/datasets/{ds_id}": {
      "delete": {
        "tags": [
          "Datasets"
        ],
        "summary": "Delete Dataset",
        "description": "Drop a dataset and every byte it owns on disk.\n\nNothing else in the app ever deletes a dataset, so a rejected or abandoned\nimport keeps its cache directory and upload copy forever.",
        "operationId": "delete_dataset_api_v1_datasets__ds_id__delete",
        "parameters": [
          {
            "name": "ds_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Ds Id"
            }
          },
          {
            "name": "user_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1,
              "title": "User Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/datasets/{ds_id}/rows/{index}": {
      "get": {
        "tags": [
          "Datasets"
        ],
        "summary": "Get Row",
        "operationId": "get_row_api_v1_datasets__ds_id__rows__index__get",
        "parameters": [
          {
            "name": "ds_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Ds Id"
            }
          },
          {
            "name": "index",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Index"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/datasets/{ds_id}/rows/{index}/columns/{column}": {
      "get": {
        "tags": [
          "Datasets"
        ],
        "summary": "Get Binary Column",
        "operationId": "get_binary_column_api_v1_datasets__ds_id__rows__index__columns__column__get",
        "parameters": [
          {
            "name": "ds_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Ds Id"
            }
          },
          {
            "name": "index",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Index"
            }
          },
          {
            "name": "column",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Column"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Body_upload_dataset_api_v1_datasets_upload_post": {
        "properties": {
          "file": {
            "type": "string",
            "contentMediaType": "application/octet-stream",
            "title": "File"
          }
        },
        "type": "object",
        "required": [
          "file"
        ],
        "title": "Body_upload_dataset_api_v1_datasets_upload_post"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  },
...
}